fix(sdk): export thor swap memo parser
