+++ .changeset/fix-sdk-thor-memo-parser-r7.md
@@ -1,0 +1,6 @@
+---
+'@vultisig/sdk': patch
+'@vultisig/cli': patch
+---
+
+Promote THOR/Maya swap-memo parsing into the public SDK API and update the CLI executor to consume the shared helper.
