+++ clients/cli/src/agent/executor.ts
@@ -7,7 +7,15 @@
  * to be flushed into the next outbound `context.recent_actions`.
  */
 import type { VaultBase, Vultisig } from '@vultisig/sdk'
-import { Chain, chainFeeCoin, getChainKind, VaultError, VaultErrorCode, Vultisig as VultisigSdk } from '@vultisig/sdk'
+import {
+  Chain,
+  chainFeeCoin,
+  getChainKind,
+  parseThorSwapMemo,
+  VaultError,
+  VaultErrorCode,
+  Vultisig as VultisigSdk,
+} from '@vultisig/sdk'
 import { formatUnits, hashTypedData, recoverAddress } from 'viem'
 
 import { VaultStateStore } from '../core/VaultStateStore'
@@ -21,60 +29,6 @@ import {
   reserveBroadcast,
 } from './broadcastJournal'
 import type { RecentAction } from './types'
-
-/**
- * THORChain swap-memo chain codes → SDK Chain enum. Used to resolve the
- * destination chain encoded in `=:CHAIN.ASSET:DEST::v0:slippage` memos
- * back to a Chain enum so vault.swap can dispatch.
- *
- * Reference: https://docs.thorchain.org/concepts/memos
- * Maya uses similar codes — additions here cover both ecosystems.
- */
-const THOR_MEMO_CHAIN_TO_ENUM: Record<string, Chain> = {
-  BTC: Chain.Bitcoin,
-  ETH: Chain.Ethereum,
-  BSC: Chain.BSC,
-  AVAX: Chain.Avalanche,
-  BASE: Chain.Base, // L2 — THORChain routinely quotes Base destinations (PR #439 review finding 1)
-  ARB: Chain.Arbitrum, // L1-via-bridge path (PR #439 review finding 1)
-  BCH: Chain.BitcoinCash,
-  LTC: Chain.Litecoin,
-  DOGE: Chain.Dogecoin,
-  GAIA: Chain.Cosmos,
-  THOR: Chain.THORChain,
-  RUNE: Chain.THORChain,
-  XRP: Chain.Ripple,
-  DASH: Chain.Dash,
-  ZEC: Chain.Zcash,
-  MAYA: Chain.MayaChain,
-  CACAO: Chain.MayaChain,
-}
-
-/**
- * THORChain abbreviated asset shortcuts → expanded `CHAIN.ASSET`. THORChain
- * memos accept both full (`XRP.XRP`) and abbreviated (`x`) notation for
- * common native assets to fit within the 250-byte memo limit when paired
- * with long destination addresses. Reference:
- * https://docs.thorchain.org/concepts/asset-notation#asset-shorthand
- */
-const THOR_MEMO_ASSET_SHORTCUTS: Record<string, string> = {
-  b: 'BTC.BTC',
-  e: 'ETH.ETH',
-  s: 'BSC.BNB',
-  a: 'AVAX.AVAX',
-  c: 'BCH.BCH',
-  l: 'LTC.LTC',
-  d: 'DOGE.DOGE',
-  g: 'GAIA.ATOM',
-  r: 'THOR.RUNE',
-  x: 'XRP.XRP',
-  cacao: 'MAYA.CACAO',
-  dash: 'DASH.DASH',
-  zec: 'ZEC.ZEC',
-  // BASE / ARB don't have documented single-letter shortcuts; THORChain
-  // emits these as the full CHAIN.ASSET form in memos. Listed in
-  // THOR_MEMO_CHAIN_TO_ENUM only.
-}
 
 // EVM chains that use nonce-based transaction ordering
 const EVM_CHAINS = new Set<string>([
@@ -915,14 +869,7 @@ export class AgentExecutor {
     const txArgs = serverTxData?.txArgs ?? {}
     const memo: string = typeof txArgs.memo === 'string' ? txArgs.memo : ''
     const parsed = parseThorSwapMemo(memo)
-
-    const toChain = THOR_MEMO_CHAIN_TO_ENUM[parsed.destChainCode]
-    if (!toChain) {
-      throw new VaultError(
-        VaultErrorCode.UnsupportedChain,
-        `signThorMsgDepositSwap: unsupported destination chain code '${parsed.destChainCode}' in memo '${memo}'.`
-      )
-    }
+    const toChain = parsed.toChain
 
     // Fund-safety: require memo destination to equal the vault's own
     // destination address. vault.swap silently substitutes the vault's
@@ -2123,73 +2070,6 @@ export function parseNonEvmEnvelope(serverTxData: any, chain: Chain): NonEvmSend
   const memo: string | undefined = typeof txArgs.memo === 'string' && txArgs.memo.length > 0 ? txArgs.memo : undefined
 
   return { chain, to, amount: amountDecimal, symbol, memo }
-}
-
-/**
- * Parsed shape of a THORChain swap memo (`=:CHAIN.ASSET:DEST[::v0:slippage]`).
- *
- * - `destChainCode` is the raw memo chain prefix (`XRP`, `ETH`, ...).
- *   Caller is responsible for mapping it to a `Chain` enum via
- *   `THOR_MEMO_CHAIN_TO_ENUM` and rejecting unsupported codes.
- * - `destAsset` is the asset ticker only — any ERC-20 contract suffix
- *   (`USDC-0X...`) is stripped because vault.swap takes the ticker.
- * - `destAddress` is the user-supplied destination on the destination
- *   chain. May be empty when the memo omits it (THORChain treats this
- *   as "refund to source"); callers should still validate against the
- *   vault's own destination address before broadcasting since vault.swap
- *   silently substitutes its own address into the broadcast memo.
- */
-export type ParsedThorSwapMemo = {
-  destChainCode: string
-  destAsset: string
-  destAddress: string
-}
-
-/**
- * Parse a THORChain swap memo into its destination-routing components.
- *
- * Accepts the shorthand notation documented at
- * https://docs.thorchain.org/concepts/asset-notation#asset-shorthand
- * (`x` → `XRP.XRP`, `e` → `ETH.ETH`, ...) — common shortcuts let memos
- * fit inside THORChain's 250-byte limit when paired with long EVM
- * destination addresses.
- *
- * Throws `VaultError(NotImplemented)` for non-swap memos (anything that
- * doesn't start with the `=:` swap prefix — e.g. LP `+:POOL` or `-:POOL`,
- * which are deferred to Phase E). Throws `VaultError(InvalidConfig)`
- * when the swap memo is structurally malformed (no CHAIN.ASSET segment).
- */
-export function parseThorSwapMemo(memo: string): ParsedThorSwapMemo {
-  if (!memo.startsWith('=:')) {
-    throw new VaultError(
-      VaultErrorCode.NotImplemented,
-      `parseThorSwapMemo: only swap memos (=:CHAIN.ASSET:DEST...) supported on this path; got memo='${memo}'. ` +
-        `LP memos (+:/-:) route through signThorMsgDepositLp; loan / validator ops out of scope.`
-    )
-  }
-
-  const memoBody = memo.slice(2) // strip leading '=:'
-  const parts = memoBody.split(':')
-
-  let chainAsset = parts[0]
-  if (chainAsset && !chainAsset.includes('.')) {
-    const expanded = THOR_MEMO_ASSET_SHORTCUTS[chainAsset.toLowerCase()]
-    if (expanded) chainAsset = expanded
-  }
-  if (!chainAsset || !chainAsset.includes('.')) {
-    throw new VaultError(
-      VaultErrorCode.InvalidConfig,
-      `parseThorSwapMemo: malformed swap memo '${memo}': missing CHAIN.ASSET in first segment.`
-    )
-  }
-
-  const [destChainCode, destAssetRaw] = chainAsset.split('.')
-  // The destAsset can carry an ERC-20 contract suffix ("ETH.USDC-0X...");
-  // for vault.swap we only need the ticker (part before `-`).
-  const destAsset = destAssetRaw?.split('-')[0] ?? ''
-  const destAddress = typeof parts[1] === 'string' ? parts[1] : ''
-
-  return { destChainCode, destAsset, destAddress }
 }
 
 /**
